apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
openpyxl==3.1.5
alembic==1.14.0
psycopg2-binary==2.9.10
